Rename class-inheritance.js to classInheritance.js
