+++ test/transform/classInheritance.js
