+++ test/transform/classTest.js
@@ -1,6 +1,6 @@
 import createTestHelpers from '../createTestHelpers';
 const {expectTransform, expectNoChange} = createTestHelpers(['class']);
-import inheritanceTests from './class-inheritance';
+import classInheritance from './classInheritance';
 
 describe('Classes', () => {
   it('should not convert functions without prototype assignment to class', () => {
@@ -230,7 +230,7 @@ describe('Classes', () => {
     );
   });
 
-  inheritanceTests();
+  classInheritance();
 
   describe('comments', () => {
     it('should preserve class comments', () => {
